Roll back TESETETH and TESTERC20AssetHolder.sol
-  wrapper methods setHoldings and setOutcomePermissoinless go into TESTAssetHolder (we will unit test transferAll against this contract)
- no longer try to  test that assets are transferred out (punt to a separate unit test)
