+++ contracts/TESTAssetHolder.sol
@@ -5,6 +5,18 @@ import './AssetHolder.sol';
 contract TESTAssetHolder is AssetHolder {
     // Public wrappers for internal methods:
 
+    function setHoldings(bytes32 channelId, uint256 amount) external {
+        holdings[channelId] = amount;
+    }
+
+    function setOutcomePermissionless(bytes32 channelId, bytes32 outcomeHash)
+        external
+        returns (bool success)
+    {
+        _setOutcome(channelId, outcomeHash);
+        return true;
+    }
+
     function isExternalAddress(bytes32 destination) public pure returns (bool) {
         return _isExternalAddress(destination);
     }
